+++ CHANGELOG.rst
@@ -53,6 +53,7 @@ Internal changes:
  - Ensure that shell files are always saved with LF linebreaks. (:issue:`547`)
  - Update the test driver to share the reference data between the different compiler versions. (:issue:`556`)
  - Add flake8-print to check usage of print function. (:issue:`566`)
+ - Add timeout for each single test. (:issue:`572`)
  - Add support for gcc-9 and clang-13 to docker tests. (:issue:`527`)
 
 5.0 (11 June 2021)